Fix method signature to work with float average

+++ src/main/java/com/utd/cs6359/assignment4/Average.java
@@ -1,8 +1,8 @@
 package com.utd.cs6359.assignment4;
 
 public class Average {
-    public int average(int k, int[] list) {
-        int average = 0;
+    public float average(int k, int[] list) {
+        float average = 0.0f;
         int n = Math.min(k, list.length);
         if (n>0) {
             for (int i=0; i<n; i++) {
